requirements.txt
flask
python-dotenv
openai
pytest
pytest-cov
